optimized display_all_menu() function
optimized by using ImGuiListClipper type. But failed to use the same method for display_category_menu(). Implemented a bool variable data_loaded and implemented checks for all display menus for further optimization. But display_category_menu() is still noticeably laggy because it is not using clipper and is rendering ALL the elements in the item.csv file.